% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{retrieve_cache}
\alias{retrieve_cache}
\alias{cache_atac_se}
\title{Download, retrieve and cache macrophage ATAC seq results}
\source{
\url{https://zenodo.org/record/1188300/files/}
}
\usage{
retrieve_cache()

cache_atac_se(verbose = FALSE)
}
\arguments{
\item{verbose}{Provide progress bar for downloading and update the cache?
(Default is FALSE)}
}
\value{
the path to the .rds file containing the SummarizedExperiment
}
\description{
Download, retrieve and cache macrophage ATAC seq results
}
\details{
This code generates the SummarizedExperiment object
used for ATAC seq peaks data used in the vignette.
}
