% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\docType{data}
\name{peaks}
\alias{peaks}
\title{Summarised ATAC-seq peaks as a GRanges object}
\format{An object of class \code{GRanges} of length 296220.}
\usage{
peaks
}
\description{
A GRanges object containing approximately 300,000 differential
accessiblity peaks estimated via limma. A description of how
this object was generated can be found in the vignette.
}
\details{
#' @format A GRanges with 296220 rows and 3 metadata columns:
\describe{
  \item{peak_id}{the ATAC peak identifer}
  \item{da_log2FC}{the estimated log2 fold change}
  \item{da_padj}{the estimated differential accessibility FDR adjusted p-value}
}
}
\keyword{datasets}
